pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "crosbi"
version = "0.1.0"
description = "CroRIS CROSBI Projekti API klijent"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "urllib3>=2.0",
    "python-dotenv>=1.0",
    "marimo>=0.8",
    "plotly>=5.18",
    "pandas>=2.1",
]

[project.scripts]
crosbi = "main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["crosbi*"]
